requirements.txt
pytorch
tqdm
argparse-dataclass
pydantic
